# apps/platform/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Flow Life Platform API
  version: 1.0.0
  description: >-
    Authoritative API for Flow Life organizational data, identity, authorization
    and workflows.


    **Implemented so far:** infrastructure health, Guardian Console session
    authentication (`login`, `logout`, `me`), capability-based authorization
    (see below), and the credential lifecycle (invitation acceptance, password
    recovery and password change), and multi-factor authentication (authenticator
    enrolment, the sign-in challenge, recovery codes and recent-verification
    step-up). No endpoint changes role assignments yet.


    ## Authorization


    What a person may do is a set of **capabilities** (for example `console.access`),
    bundled into roles that are assigned to people. Capabilities are named by
    identifier; **role names are never exposed** and must never be used to decide what
    to show or do. The platform evaluates capabilities against the person's *current*
    assignments on every request: a revoked role takes effect on the next request with no
    new sign-in, and nothing about them is stored in the session or accepted from the
    client. A signed-in account without a capability receives `403`.


    ## Session authentication (Guardian Console)


    The Console signs in with a same-origin, host-only session cookie (ADR 0016).
    There is no bearer token and no CORS: the Console and the API share one origin.


...
      differ from production.

...

...

...

...

...
servers:
...
paths:
...
  /mfa/recovery-codes:
    post:
      operationId: regenerateRecoveryCodes
      summary: Replace the recovery codes
      description: >-
        Replaces the signed-in account's recovery codes with a new set, returned **once**. Fresh proof
        is required in the body: the **current password and a second factor** (`code` or
        `recovery_code`); a session alone is never enough. Every earlier code, used or not, stops
        working in the same transaction. On success the session is rotated (new id and CSRF token)
        and recorded as freshly verified. Rate limited per account and per source address, so a
        stolen session cannot be used to guess the password or a code.
      tags: [Authentication]
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/XsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecurityProofRequest'
      responses:
        '200':
          description: The new recovery codes. Shown once.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoveryCodes'
        '401':
          description: Not signed in, the session has expired, or the account can no longer sign in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '419':
          description: Missing or invalid CSRF token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '422':
          description: Nothing was changed. `errors.current_password`, `errors.code` or `errors.recovery_code`, or a malformed request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '429':
          description: Too many attempts for this account or from this address.
          headers:
            Retry-After:
              description: Seconds until another attempt is accepted.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /mfa/authenticator:
    post:
      operationId: beginAuthenticatorReplacement
      summary: Start replacing the authenticator
      description: >-
        Starts replacing the signed-in account's authenticator (a new phone, say). Fresh proof in the
        body: the **current password and a second factor**. Returns a **new** secret and provisioning
        URI **once**, as *pending*: the authenticator in use **keeps working** until the new one is
        proved with `POST /mfa/authenticator/confirm`, so an abandoned or failed replacement cannot
        strand the account. Asking again replaces the pending secret. On success the session is
        rotated and recorded as freshly verified. Not cacheable.
      tags: [Authentication]
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/XsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecurityProofRequest'
      responses:
        '200':
          description: The new secret and provisioning URI. Shown once; the old authenticator still works.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticatorSetup'
        '401':
          description: Not signed in, the session has expired, or the account can no longer sign in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '419':
          description: Missing or invalid CSRF token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '422':
          description: Nothing was changed. `errors.current_password`, `errors.code` or `errors.recovery_code`, or a malformed request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '429':
          description: Too many attempts for this account or from this address.
          headers:
            Retry-After:
              description: Seconds until another attempt is accepted.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-flowlife-session
      description: >-
        Same-origin, host-only session cookie (ADR 0016). Not readable by scripts.
        Bounded by 30 minutes of request inactivity and a 12-hour absolute lifetime.
  parameters:
    XsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: false
      description: >-
        The value of the `XSRF-TOKEN` cookie. Required for state-changing requests
        unless the browser marks the request `Sec-Fetch-Site: same-origin`.
      schema:
        type: string
  schemas:
...
    Message:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ValidationError:
      type: object
      required: [message, errors]
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    SecurityProofRequest:
      type: object
      description: Fresh proof. The current password and exactly one of `code` and `recovery_code`.
      required: [current_password]
      properties:
        current_password:
          type: string
          maxLength: 1024
          format: password
          writeOnly: true
          description: The account's current password, re-proved. Secret.
        code:
          type: string
          maxLength: 32
          writeOnly: true
          description: A 6-digit code from the authenticator app (spaces are ignored). Secret.
        recovery_code:
          type: string
          maxLength: 64
          writeOnly: true
          description: One of the account's single-use recovery codes. Secret.
    AuthenticatorSetup:
      type: object
      description: >-
        What is needed to add an authenticator. Shown **once**, and never retrievable again: the server
        keeps only the encrypted secret. Secret.
      required: [secret, otpauth_uri]
      properties:
        secret:
          type: string
          readOnly: true
          description: The base32 secret, for typing into an authenticator app by hand.
        otpauth_uri:
          type: string
          readOnly: true
          description: The `otpauth://` provisioning URI. Render it as a QR code **in the browser**; never send it to a QR-code service. Contains the secret.
        expires_at:
          type: string
          format: date-time
          description: For a sign-in enrolment, when the pending sign-in ends.
    RecoveryCodes:
      type: object
      description: New recovery codes, shown **once** and stored only as digests. Secret.
      required: [recovery_codes]
      properties:
        recovery_codes:
          type: array
          readOnly: true
          items:
            type: string
          description: Ten single-use codes, grouped for reading (`ABCD-EFGH-JKMN-PQRS`).
...